Updates and bugfixes

+++ src/main/java/com/mcmoddev/lib/item/ItemBow.java
@@ -31,15 +31,18 @@ public class ItemBow extends net.minecraft.item.ItemBow {
 	 * Called when the player stops using an Item (stops holding the right mouse button).
 	 */
 	@Override
-	public void onPlayerStoppedUsing(final ItemStack stack, final World worldIn, final EntityLivingBase entityLiving, final int timeLeft) {
+	public void onPlayerStoppedUsing(final ItemStack stack, final World worldIn,
+			final EntityLivingBase entityLiving, final int timeLeft) {
 		if (entityLiving instanceof EntityPlayer) {
 			final EntityPlayer entityPlayer = (EntityPlayer) entityLiving;
-			final boolean flag = entityPlayer.capabilities.isCreativeMode || (EnchantmentHelper.getEnchantmentLevel(Enchantments.INFINITY, stack) > 0);
+			final boolean flag = entityPlayer.capabilities.isCreativeMode
+					|| (EnchantmentHelper.getEnchantmentLevel(Enchantments.INFINITY, stack) > 0);
 			// Check for specific Arrows
 			ItemStack itemStack = this.myFindAmmo(entityPlayer);
 
 			int i = this.getMaxItemUseDuration(stack) - timeLeft;
-			i = net.minecraftforge.event.ForgeEventFactory.onArrowLoose(stack, worldIn, entityPlayer, i, (itemStack != null) || flag);
+			i = net.minecraftforge.event.ForgeEventFactory.onArrowLoose(stack, worldIn,
+					entityPlayer, i, (itemStack != null) || flag);
 			if (i < 0) {
 				return;
 			}
@@ -51,28 +54,39 @@ public class ItemBow extends net.minecraft.item.ItemBow {
 
 				final float f = getArrowVelocity(i);
 
-				if ((double) f >= 0.1D) {
+				if (f >= 0.1D) {
 					// OLD
-					//final boolean flag1 = flag && (itemstack.getItem() instanceof ItemArrow); // Forge: Fix consuming custom arrows.
+					// final boolean flag1 = flag && (itemstack.getItem() instanceof ItemArrow); //
+					// Forge: Fix consuming custom arrows.
 					// NEW
-					final boolean flag1 = entityPlayer.capabilities.isCreativeMode || (itemStack.getItem() instanceof ItemArrow ? ((ItemArrow) itemStack.getItem()).isInfinite(itemStack, stack, entityPlayer) : false);
+					final boolean flag1 = entityPlayer.capabilities.isCreativeMode
+							|| (itemStack.getItem() instanceof ItemArrow
+									? ((ItemArrow) itemStack.getItem()).isInfinite(itemStack, stack,
+											entityPlayer)
+									: false);
 
 					if (!worldIn.isRemote) {
-						final ItemArrow itemArrow = ((ItemArrow) (itemStack.getItem() instanceof ItemArrow ? itemStack.getItem() : Items.ARROW));
-						final EntityArrow entityArrow = itemArrow.createArrow(worldIn, itemStack, entityPlayer);
-						entityArrow.setAim(entityPlayer, entityPlayer.rotationPitch, entityPlayer.rotationYaw, 0.0F, f * 3.0F, 1.0F);
+						final ItemArrow itemArrow = ((ItemArrow) (itemStack
+								.getItem() instanceof ItemArrow ? itemStack.getItem()
+										: Items.ARROW));
+						final EntityArrow entityArrow = itemArrow.createArrow(worldIn, itemStack,
+								entityPlayer);
+						entityArrow.setAim(entityPlayer, entityPlayer.rotationPitch,
+								entityPlayer.rotationYaw, 0.0F, f * 3.0F, 1.0F);
 
 						if (f == 1.0F) {
 							entityArrow.setIsCritical(true);
 						}
 
-						final int j = EnchantmentHelper.getEnchantmentLevel(Enchantments.POWER, stack);
+						final int j = EnchantmentHelper.getEnchantmentLevel(Enchantments.POWER,
+								stack);
 
 						if (j > 0) {
-							entityArrow.setDamage(entityArrow.getDamage() + ((double) j * 0.5D) + 0.5D);
+							entityArrow.setDamage(entityArrow.getDamage() + (j * 0.5D) + 0.5D);
 						}
 
-						final int k = EnchantmentHelper.getEnchantmentLevel(Enchantments.PUNCH, stack);
+						final int k = EnchantmentHelper.getEnchantmentLevel(Enchantments.PUNCH,
+								stack);
 
 						if (k > 0) {
 							entityArrow.setKnockbackStrength(k);
@@ -91,7 +105,10 @@ public class ItemBow extends net.minecraft.item.ItemBow {
 						worldIn.spawnEntity(entityArrow);
 					}
 
-					worldIn.playSound((EntityPlayer) null, entityPlayer.posX, entityPlayer.posY, entityPlayer.posZ, SoundEvents.ENTITY_ARROW_SHOOT, SoundCategory.NEUTRAL, 1.0F, (1.0F / ((itemRand.nextFloat() * 0.4F) + 1.2F)) + (f * 0.5F));
+					worldIn.playSound((EntityPlayer) null, entityPlayer.posX, entityPlayer.posY,
+							entityPlayer.posZ, SoundEvents.ENTITY_ARROW_SHOOT,
+							SoundCategory.NEUTRAL, 1.0F,
+							(1.0F / ((itemRand.nextFloat() * 0.4F) + 1.2F)) + (f * 0.5F));
 
 					if (!flag1) {
 						--itemStack.stackSize;
@@ -137,7 +154,9 @@ public class ItemBow extends net.minecraft.item.ItemBow {
 	 * Called when the equipped item is right clicked.
 	 */
 	@Override
-	public ActionResult<ItemStack> onItemRightClick(final ItemStack itemStackIn, final World worldIn, final EntityPlayer playerIn, final EnumHand handIn) {
+	public ActionResult<ItemStack> onItemRightClick(final ItemStack itemStackIn,
+			final World worldIn, final EntityPlayer playerIn, final EnumHand handIn) {
+		// Check for specific Arrows
 		final boolean flag = this.myFindAmmo(playerIn) != null;
 
 		if (!playerIn.capabilities.isCreativeMode && !flag) {
